Initial commit
This commit migrates the project from our internal git to the public repository.
We decided not to keep the git history because it contains some possibly sensitive internal information.
Here are some fun facts about the history at the point of this commit though:
Number of commits: 1867
Date of initial commit: 2024-07-02