ci: codespell: add workflow for spell checking
The codespell[1] project is a tool that prevents common spelling mistakes.
It achieves a low false-positive rate (which is required e.g. for use in
CI systems) by operating on a list of known common typos instead of known
correct spellings.

Run the workflow on `pull_request` and `push` to the `main` branch,
to check for correct spelling before and after merging.

Also check spelling on a regular basis (20:48 on thursdays) because new
`codespell` versions with new dictionaries are release regularly and newly
added corrections should not be miss-attributed to the next new pull
request.

Also allow manually running the workflow by using `workflow_dispatch`.

[1]: https://github.com/codespell-project/codespell