README: replace web demo URL with GitHub pages variant
A recent commit has added support for building the example project and
deploying them using GitHub pages.
This means an up do date version of the example project is available
nearly immediately after e.g. merging a pull request into `main`.
The status of this deployment is also much more transparent to an outside
observer, because the action runs are public.

Use the GitHub pages backed version of the example project as the official
web demo.