ci: build: add job to build the web version of the example project
A build job can be a valuable resource for newcomers to a project to learn
how to build the software.

Add a build job as an example of how to build a werkbank project and as an
integration test to check if everything still compiles.

Run the build for pull requests (to check for regressions introduced by
them) and for pushes to `main` (to check for regressions introduced by
merge conflicts and as a source for up to date artifacts).